+++ src/index.ts
@@ -1,5 +1,4 @@
 export * from './client';
-export * as client from './client';
 export * from './constants';
 export * from './types';
 export * from './utils';
